refactor: moved files to ChatRoomManager folder, created a parent component for room list and room selector and moved the logic from chat context to the parent component
